Extract interface for StoreDataManagerQuery and move impl to memory data manager
Some data managers may not work with all ArtifactStores in memory, or
may be more efficient working in some backing store like a database or
clustered Infinispan. In these cases, we need to make room for new
implementations of StoreDataManagerQuery that can take full advantage
of the native storage service.

This change refactors StoreDataManagerQuery to MemoryStoreDataManagerQuery
(specific to MemoryStoreDataManager and its derivatives), and then re-extracts
the StoreDataManagerQuery interface.
